Dev auth vadim

+++ src/main/java/com/example/rest/UrlShortenerController.java
@@ -8,6 +8,7 @@ import com.example.data.url_view.UrlViewDto;
 import com.example.exceptions.UrlExpiredException;
 import com.example.exceptions.UrlNotFoundException;
 import com.example.exceptions.UserNotFoundException;
+import com.example.url_profile.UrlProfile;
 import com.example.url_profile.UrlProfileService;
 import com.example.url_profile.UrlProfileValidationService;
 import com.example.url_view.UrlViewMapper;
@@ -16,17 +17,22 @@ import com.example.user.User;
 import com.example.user.UserService;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.view.RedirectView;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @RestController
-@RequestMapping("v1/urls")
+@RequestMapping("V1/urls")
 public class UrlShortenerController {
     private UrlProfileService urlProfileService;
     private UserService userService;
@@ -49,6 +55,9 @@ public class UrlShortenerController {
         if (!urlProfileValidationService.isValidURL(urlProfileRequest.getFullUrl())) {
             return ResponseEntity.badRequest().body("Invalid URL");
         }
+//        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+//        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+//        String username = userDetails.getUsername();
 
         LocalDateTime now = LocalDateTime.now();
 
@@ -79,18 +88,26 @@ public class UrlShortenerController {
 
     @GetMapping("")
     public ResponseEntity<List<UrlProfileView>> getAllUrlProfiles() {
-        List<UrlProfileResponse> urlProfileResponses = urlProfileService.getAllUrls();
-        List<UrlProfileView> urlProfileViews = new ArrayList<>();
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+        String username = userDetails.getUsername();
+        User user = userService.getUserByUsername(username);
+        Set<UrlProfile> urls = user.getUrls();
 
-        for(UrlProfileResponse urlProfileResponse : urlProfileResponses){
+        List<UrlProfileView> urlProfileViews = new ArrayList<>();
+        for (UrlProfile urlProfile : urls) {
             urlProfileViews.add(new UrlProfileView(
-                    urlProfileResponse.getLongUrl(),
-                    urlProfileResponse.getShortUrl(),
-                    urlProfileResponse.getCreatedAt(),
-                    urlProfileResponse.getValidTo(),
-                    urlProfileResponse.getUser().getUsername(),
-                    urlViewService.getCountOfRedirectsForUrl(urlProfileResponse.getId()),
-                    urlViewService.getViewsForUrl(urlProfileResponse.getId()).stream().map(urlViewMapper::fromUrlViewResponseToDto).collect(Collectors.toList())));
+                    urlProfile.getLongUrl(),
+                    urlProfile.getShortUrl(),
+                    urlProfile.getCreatedAt(),
+                    urlProfile.getValidTo(),
+                    urlProfile.getUser().getUsername(),
+                    urlViewService.getCountOfRedirectsForUrl(urlProfile.getId()),
+                    urlViewService.getViewsForUrl(urlProfile.getId())
+                            .stream()
+                            .map(urlViewMapper::fromUrlViewResponseToDto)
+                            .collect(Collectors.toList())
+            ));
         }
 
         return ResponseEntity
@@ -140,6 +157,16 @@ public class UrlShortenerController {
 
     @DeleteMapping("/{shortUrl}")
     public ResponseEntity<?> deleteByUrl(@PathVariable String shortUrl) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+        String username = userDetails.getUsername();
+        User user = userService.getUserByUsername(username);
+
+        UrlProfile urlProfile = urlProfileService.getUrlProfileByShortUrl(shortUrl);
+        boolean isUserUrl = user.getUrls().stream().anyMatch(profile -> profile.getShortUrl().equals(shortUrl));
+        if (!isUserUrl) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Url not found");
+        }
         urlProfileService.deleteByShortUrl(shortUrl);
         return ResponseEntity
                 .ok()
